// src/app/pages/collections/phantasia/tools/shared/interfaces.ts
export interface CDCase {
  id: number;
  title: string;
  artist: string;
  imageUrl: string;
  model: THREE.Group;
  isFlipped: boolean;
  position: THREE.Vector3;
  rotation: THREE.Euler;
  mixer?: THREE.AnimationMixer;
  animations?: THREE.AnimationClip[];
  openAction?: THREE.AnimationAction;
  closeAction?: THREE.AnimationAction;
  armature?: THREE.Object3D | null;
  isOpen: boolean;
  glowMaterial?: THREE.MeshBasicMaterial;
  carouselIndex: number;
  targetPosition: THREE.Vector3;
  isActive: boolean;
  isDeactivating: boolean;
  currentLerpAlpha: number;
  initialPosition: THREE.Vector3;
}